BackEnd/requirements.txt
numpy==1.20.1
mysql-connector-python==8.0.23
opencv-contrib-python-headless==4.4.0.46